Move selection reporting functions and types to `CoinSelection`.
We eventually want to include information about selected collateral in
these reports, so it makes more sense to locate the reports in the
top-level `CoinSelection` module.
